Cleaning up route params handling
