+++ public/jobIcons/cj/1/archer.png
