+++ public/jobIcons/cj/1/armorer.png
