+++ public/jobIcons/cj/1/bard.png
