+++ public/jobIcons/cj/1/conjurer.png
