+++ public/jobIcons/cj/1/monk.png
