+++ public/jobIcons/cj/1/pugilist.png
