+++ public/jobIcons/cj/1/warrior.png
